% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_algorithm.R
\name{find_leaf1}
\alias{find_leaf1}
\title{Find Leaf Nodes}
\usage{
find_leaf1(tree)
}
\arguments{
\item{tree}{A tibble representing the decision tree.}
}
\value{
A vector of node identifiers corresponding to leaf nodes.
}
\description{
Helper function to identify leaf nodes in a decision tree.
The function filters the nodes where `name` is "leaf" and returns the corresponding node identifiers.
}
