% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_algorithm.R
\name{greedy_cart}
\alias{greedy_cart}
\title{Greedy Algorithm}
\usage{
greedy_cart(
  x,
  y,
  data,
  type = NULL,
  num_leaf = NULL,
  depth = NULL,
  num_split = 2,
  min_num = 1,
  m = 0,
  unique = FALSE
)
}
\arguments{
\item{x}{Column/list name(s) of the x value(s).}

\item{y}{Column/list name of the y value.}

\item{data}{Tibble or named list with data.}

\item{type}{"reg" for regression tree, "class" for classification tree. If `type` is missing, the function tries to guess the type.}

\item{num_leaf}{Condition to end: the tree has `num_leaf` leaves. Must be greater than or equal to 1. The default value is the maximum number of leaves (the number of data points).}

\item{depth}{Condition to end: the tree has depth `depth`. Must be greater than or equal to 0. The default value is the maximum achievable depth.}

\item{num_split}{Split only nodes which contain at least `num_split` elements. Must be greater than or equal to 2.}

\item{min_num}{Only split a node if both child nodes have at least `min_num` elements. Must be greater than or equal to 1.}

\item{m}{Parameter for Random Forest algorithm: positive number of coordinates used in each iteration. The default value is the dimension of the data.}

\item{unique}{Parameter for classification data: if `unique` is set to TRUE, nodes where all data points in this node have the same class (y value) are not split. The default value is FALSE.}
}
\value{
An environment containing the elements `dim`, `values`, and `tree`.
\itemize{
  \item `dim`: The dimension of the data.
  \item `values`: Data in a tibble format.
  \item `tree`: Decision tree in the form of a tibble. It has the following columns:
    \itemize{
      \item `node`: Node identifier, where node n has child nodes 2n and 2n + 1.
      \item `name`: Type of the node (root, inner node, leaf).
      \item `split_index`: Index at which data is split.
      \item `split_point`: Value at which data is split.
      \item `y`: y values of data contained in this node.
      \item `A`: x values of data contained in this node.
      \item `c_value`: The approximate value for the data elements in a node.
    }
}
}
\description{
Greedy algorithm for either regression or classification data.
}
\examples{
X <- runif(100,0,1)
e <- rnorm(100,0,0.2)
Y <- sin(2*pi*X) + e
data <- list(a = X, b = Y)
val <- greedy_cart(x = a, y = b, data = data, type = "reg")
val$values
val$tree

X1 <- runif(200,0,1)
X2 <- runif(200,0,1)
e <- rnorm(200,0,0.05)
k <- function(x,y) (x-0.5)*(y-0.5)
g <- function(x,y,e){
  Y <- c()
  for(i in seq_along(x)){
   if(k(X1[i],X2[i]) - e[i] <= 0){
     Y[i] <- 1
    } else{
      Y[i] <- 2
    }
  }
  Y
}
tbl <- tibble(x1 = X1, x2 = X2, y = g(X1,X2,e))
val <- greedy_cart(x = c(x1,x2), y = y, data = tbl, type = "class", depth = 3)
val$values
val$tree
}
