% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_algorithm.R
\name{greedy_cart_classification}
\alias{greedy_cart_classification}
\title{Greedy Algorithm (Classification)}
\usage{
greedy_cart_classification(
  data,
  num_leaf = NULL,
  depth = NULL,
  num_split = 2,
  min_num = 1,
  m = 0,
  unique = FALSE
)
}
\arguments{
\item{data}{A named list containing classification data. The x values have the name `x` and are in the form of a matrix where the row number gives the dimension of the data. The y values have the name `y` and are in the form of a vector.}

\item{num_leaf}{Condition to end: the tree has `num_leaf` leaves. Must be greater than or equal to 1. The default value is the maximum number of leaves (the number of data points).}

\item{depth}{Condition to end: the tree has depth `depth`. Must be greater than or equal to 0. The default value is the maximum achievable depth.}

\item{num_split}{Split only nodes which contain at least `num_split` elements. Must be greater than or equal to 2.}

\item{min_num}{Only split a node if both child nodes have at least `min_num` elements. Must be greater than or equal to 1.}

\item{m}{Parameter for Random Forest algorithm: positive number of coordinates used in each iteration. The default value is the dimension of the data.}

\item{unique}{If `unique` is set to TRUE, nodes where all data points in this node have the same class (y value) are not split. The default value is FALSE.}
}
\value{
An environment containing the elements `dim`, `values`, and `tree`.
\itemize{
  \item `dim`: The dimension of the data.
  \item `values`: Data in a tibble format.
  \item `tree`: Decision tree in the form of a tibble. It has the following columns:
    \itemize{
      \item `node`: Node identifier, where node n has child nodes 2n and 2n + 1.
      \item `name`: Type of the node (root, inner node, leaf).
      \item `split_index`: Index at which data is split.
      \item `split_point`: Value at which data is split.
      \item `y`: y values of data contained in this node.
      \item `A`: x values of data contained in this node.
      \item `c_value`: The approximate value for the data elements in a node.
    }
}
}
\description{
Greedy algorithm for classification data.
}
\examples{
X1 <- runif(200,0,1)
X2 <- runif(200,0,1)
e <- rnorm(200,0,0.2)
kappa <- function(x,y) y - 0.5 - 0.3*sin(2*pi*x)
f <- function(x,y,e){
  Y <- c()
  for(i in seq_along(x)){
    if(kappa(X1[i],X2[i]) - e[i] <= 0){
      Y[i] <- 1
    } else{
      Y[i] <- 2
    }
  }
  Y
}
data <- list(x = matrix(c(X1,X2), nrow = 2, byrow = TRUE), y = f(X1,X2,e))
val <- greedy_cart_classification(data, num_split = 10)
val$values
val$tree
}
