% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction}
\alias{prediction}
\title{Prediction of datapoint or a set of data points in a tree or a set of trees}
\usage{
prediction(list_tree, list_x, type = NULL)
}
\arguments{
\item{list_tree}{a list containing an arbitrary number greater or equal to 1 of trees in tibble form\cr 
as used by `greedy_cart()`, `bagging()` or `random_forest()`}

\item{list_x}{a matrix of values to be predicted, where each column corresponds to a datapoint \cr
the number of rows corresponds to the dimension of the x-values in the list/matrix}

\item{type}{“reg” for regression tree\cr
“cla” for classification tree\cr}
}
\value{
A numeric vector containing the predicted values. For regression trees (`type = "reg"`), 
the predictions are the mean of the predictions across all trees for each data point. 
For classification trees (`type = "cla"`), the predictions are the most frequent class label across all trees for each data point.
}
\description{
Prediction of datapoint or a set of data points in a tree or a set of trees
}
\examples{
X1 <- runif(100, 0, 1)
X2 <- runif(100, 0, 1)
e <- rnorm(100, 0, 0.1)
Y <- X1^2 + X2 + e
data_reg_li <- list(a = X1, b=X2, y= Y)
list_tree <- random_forest(x = c(X1,X2), y = Y, data = data_reg_li, type = "reg", B = 5, A = 10, m = 1)
list_x <- matrix(c(0, 0.2, 0.3, 0.4, 0.6, 0.7, 0.9, 1), nrow = 2)
predictions <- prediction(list_tree, list_x, type = "reg")
}
