% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{random_forest_classification}
\alias{random_forest_classification}
\title{Random Forest Algorithm for classification}
\usage{
random_forest_classification(
  data,
  B,
  A = NULL,
  m = 0,
  num_leaf = NULL,
  depth = NULL,
  num_split = 2,
  min_num = 1,
  unique = F
)
}
\arguments{
\item{data}{a named list with classification data\cr 
the x-values have the name x and are in the form of a matrix\cr
the row-number indicating the dimension of the data\cr
the y-values have name y and are in the form of a vector\cr
the row-number must be the same as that of x}

\item{B}{number of bootstrap samples to be created}

\item{A}{sample size to be used\cr
must be greater than 0 and less than or equal to the number of observations\cr
the default value corresponds to the total size of data}

\item{m}{positive number of coordinates used in each iteration\cr
the default value is the dimension of the data}

\item{num_leaf}{Termination condition: The tree has `num_leaf` leaves.\cr
Must be greater than or equal to 1\cr
The default value is the number of data-points}

\item{depth}{Termination condition: The tree has depth `depth`\cr
The default value is the maximum achievable depth}

\item{num_split}{only split nodes that contain at least `num_split` elements\cr
must be greater than or equal to 2}

\item{min_num}{only divides a nod if both child nodes at least `min_num` elements\cr
must be greater than or equal to 1}

\item{unique}{if `unique`is set to TRUE, nodes where all data points in this node have the the class are not split\cr
The default value is FALSE}
}
\value{
a list of trees `B`
}
\description{
Random Forest Algorithm for classification
}
\examples{
X1 <- runif(100, 0, 1)
X2 <- runif(100, 0, 1)
e <- rnorm(100, 0, 0.1)
linear_comb <- 1.8 * X1 + 0.3 * X2 + e
Y <- ifelse(linear_comb > 1, 1, 2)
data <- list(x = matrix(c(X1, X2), nrow = 2, byrow = TRUE), y = Y)
random_forest_classification(data, B = 5, A = 10, m = 1, num_leaf = 10, depth = NULL, num_split = 2, min_num = 1, unique = F)
}
