% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{random_forest_regression}
\alias{random_forest_regression}
\title{Random Forest Algorithm}
\usage{
random_forest_regression(
  data,
  B,
  A = NULL,
  m = 0,
  num_leaf = NULL,
  depth = NULL,
  num_split = 2,
  min_num = 1
)
}
\arguments{
\item{data}{a named list with regression data\cr 
the x-values have the name x and are in the form of a matrix\cr
the row-number indicating the dimension of the data\cr
the y-values have name y and are in the form of a vector\cr
the row-number must be the same as that of x}

\item{B}{number of bootstrap samples to be created}

\item{A}{sample size to be used\cr
must be greater than 0 and less than or equal to the number of observations\cr
the default value corresponds to the total size of data}

\item{m}{positive number of coordinates used in each iteration\cr
the default value is the dimension of the data}

\item{num_leaf}{Termination condition: The tree has `num_leaf` leaves.\cr
Must be greater than or equal to 1\cr
The default value is the number of data-points}

\item{depth}{Termination condition: The tree has depth `depth`\cr
The default value is the maximum achievable depth}

\item{num_split}{only split nodes that contain at least `num_split` elements\cr
must be greater than or equal to 2}

\item{min_num}{only divides a nod if both child nodes at least `min_num` elements\cr
must be greater than or equal to 1}
}
\value{
a list of trees `B`
}
\description{
Random Forest Algorithm for regression
}
\examples{
X1 <- runif(100, 0, 1)
X2 <- runif(100, 0, 1)
e <- rnorm(100, 0, 0.1)
Y <- X1^2 + X2+ e
data <- list(x = matrix(c(X1, X2), nrow = 2, byrow = TRUE), y = Y)
random_forest_regression(data, B = 10, A = 20, m = 1, num_leaf = 10, depth = NULL, num_split = 2, min_num = 1)
}
